refactor(management): split validators/mod.rs along metric-definition / experiment-type boundary (Closes #585)
- Extract MetricLookup trait + impl, validate_metric_definition, all per-type
  helpers (filtered_mean, windowed_count, composite, metricql_arm,
  diagnostics_to_status, store_err_to_status, identifier_re) into
  validators/metric_definition.rs (~1100 lines incl. tests).
- Extract validate_meta/switchback/quasi/equivalence_test_config into
  validators/experiment_types.rs (~460 lines incl. tests).
- Reduce validators/mod.rs to the validate_starting() dispatcher + pub mod
  declarations + re-exports (~53 lines).
- cycle.rs, filter_sql.rs, metricql/ untouched.
- All 335 unit tests pass; public API surface (validate_starting,
  validate_metric_definition, MetricLookup, validate_equivalence_test_config,
  metricql::*) byte-identical via re-exports in mod.rs.
